# doc/swagger/menu.yaml
...
swagger: '2.0'
...
info:
  version: "1.0.0"
  title: 菜单接口

#host: 127.0.0.1:7443
#basePath: /v1
#schemes:
#  - https
#produces:
#  - application/json
#consumes:
#  - application/json

# Describe your paths here
paths:
...
  /menu/page:
    get:
      summary: 分页查询菜单
      description: |
        分页查询菜单
      operationId: getByPage
      parameters:
        - name: start
          in: query
          description: 分页开始索引
          type: integer
          format: int32
        - name: length
          in: query
          description: 每页数量
          type: integer
          format: int32
        - name: name
          in: query
          description: 菜单名称
          type: string
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
  /menu/status:
    get:
      summary: 通过状态查询
      description: |
        通过状态查询
      operationId: getByStatus
      parameters:
        - name: status
          in: query
          description: 状态（0，不显示，1，显示，2，显示热门）
          type: integer
          format: int32
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
  /menu/name:
    get:
      summary: 通过名称查询
      description: |
        通过名称查询
      operationId: getByName
      parameters:
        - name: name
          in: query
          description: 菜单名称
          type: string
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/responseModel'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
definitions:
...
  responseModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      data:
        type: string
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /garen.org/oss/1.0.0
schemes:
 - https
 - http